Require root

+++ ack-watch.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+if ! [ $(id -u) = 0 ]; then
+   echo "The script need to be run as root." >&2
+   exit 1
+fi
+
 BREAK="false"
 trap 'trap - SIGTERM && export BREAK="true"; kill -- -$$;' SIGINT SIGTERM EXIT
